moved liveview router into routes folder, makes more sense
