+++ routes/resolution.js
@@ -7,7 +7,7 @@ const databaseInterface = require("../lib/database");
 const tokenProcessor = require("../lib/token");
 const resUtil = require("../lib/resUtil");
 const routingUtil = require("../lib/routingUtil");
-const liveView = require("../lib/liveView").router;
+const liveView = require("./liveview").router;
 const credentials = require("../lib/credentials");
 
 const issueError = resUtil.issueError;
